Moved `getImageSrc` to a lib file
